feat: add auth & transaction CRUD use cases, fix imports & types
Add authentication use cases: login with Google and logout, with corresponding unit tests
Implement create, edit, and delete transaction use cases including their interfaces, DTOs, and full test suites
Update all presentation hooks to use the correct usecase import paths (migrated from commands directory)
Fix list transactions handler to use TransactionType enum instead of hardcoded string literals
