+++ src/application/usecases/auth/logout.handler.ts
@@ -1,0 +1,5 @@
+import { authRepository } from '../../../infrastructure/supabase/auth.repository';
+
+export async function logoutHandler(): Promise<void> {
+  await authRepository.signOut();
+}
