+++ src/presentation/hooks/use-delete-transaction.test.ts
@@ -1,6 +1,6 @@
 import { renderHook, act } from '@testing-library/react';
 import { useDeleteTransaction } from './use-delete-transaction';
-import { DeleteTransactionUseCase } from '@/application/commands/transaction/delete-transaction/delete-transaction.usecase';
+import { DeleteTransactionUseCase } from '@/application/usecases/transaction/delete-transaction/delete-transaction.usecase';
 
 jest.mock('@/application/commands/transaction/delete-transaction/delete-transaction.usecase');
